server/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
PyPDF2
faiss-cpu
transformers
torch
sentence-transformers
python-dotenv
pydantic
